+++ support/src/main/java/pt/estga/support/services/ContactRequestService.java
@@ -1,45 +1,60 @@
 package pt.estga.support.services;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import pt.estga.support.repositories.ContactRequestRepository;
-import pt.estga.support.mappers.ContactRequestMapper;
-import pt.estga.support.enums.ContactStatus;
 import pt.estga.support.dtos.ContactRequestDto;
 import pt.estga.support.entities.ContactRequest;
+import pt.estga.support.enums.ContactStatus;
+import pt.estga.support.mappers.ContactRequestMapper;
+import pt.estga.support.repositories.ContactRequestRepository;
 import pt.estga.sharedweb.exceptions.ResourceNotFoundException;
+import pt.estga.sharedweb.filtering.QueryProcessor;
+import pt.estga.sharedweb.models.PagedRequest;
+import pt.estga.sharedweb.models.QueryResult;
 
 import java.time.Instant;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
-@Transactional
-public class ContactRequestCommandService {
+@Transactional(readOnly = true)
+public class ContactRequestService {
 
     private final ContactRequestRepository repository;
     private final ContactRequestMapper mapper;
+    private final QueryProcessor<ContactRequest> queryProcessor;
 
+    @Transactional
     public ContactRequest create(ContactRequestDto dto) {
         ContactRequest contact = mapper.toEntity(dto);
         contact.setStatus(ContactStatus.PENDING);
         contact.setCreatedAt(Instant.now());
-
         return repository.save(contact);
     }
 
+    public Optional<ContactRequest> findById(Long id) {
+        return repository.findById(id);
+    }
+
+    public Page<ContactRequest> search(PagedRequest request) {
+        QueryResult<ContactRequest> result = queryProcessor.process(request);
+        return repository.findAll(result.specification(), result.pageable());
+    }
+
+    @Transactional
     public ContactRequest updateStatus(Long id, ContactStatus status) {
         ContactRequest contact = repository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Contact request not found with id: " + id));
-
         contact.setStatus(status);
         return repository.save(contact);
     }
 
+    @Transactional
     public void delete(Long id) {
         ContactRequest contact = repository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Contact request not found with id: " + id));
-
         repository.softDelete(contact);
     }
 }
